added user authorization to the front end

+++ src/components/Main.js
@@ -7,45 +7,82 @@ function Main(props) {
 
     const [people, setPeople] = useState([]);
 
-    const URL = "https://express-react-week13day3.herokuapp.com/people/";
-
-    const getPeople = async () => {
-        const response = await fetch(URL)
-        const data = await response.json()
-        setPeople(data)
-    }
+    // const URL = "https://express-react-week13day3.herokuapp.com/people/";
+    const URL = "http://localhost:3001/people/";
 
     const createPeople = async (person) => {
-        await fetch(URL, {
+        if(!props.user) return;
+        const token = await props.user.getIdToken();
+        const response = await fetch(URL, {
             method: "POST",
             headers: {
                 "Content-Type": "Application/json",
+                "Authorization": "Bearer " + token
             },
             body: JSON.stringify(person),
         })
 
-        getPeople()
+        const people = await response.json();
+        setPeople(people);
+        
     }
 
     const updatePeople = async (person, id) => {
-        await fetch(URL + id, {
+        if(!props.user) return;
+        const token = await props.user.getIdToken();
+        const response = await fetch(URL + id, {
             method: 'PUT',
             headers: {
                 "Content-Type": "Application/json",
+                "Authorization": "Bearer " + token
             },
             body: JSON.stringify(person),
         });
-        getPeople();
+
+        const people = await response.json();
+        setPeople(people);
     }
 
     const deletePeople = async (id) => {
-        await fetch(URL+id, {
+        if(!props.user) return;
+        const token = await props.user.getIdToken();
+        const response = await fetch(URL+id, {
             method: "DELETE",
+            headers: {
+                "Authorization": "Bearer " + token
+            }
         });
-        getPeople();
+        
+        const people = await response.json();
+        setPeople(people);
     };
 
-    useEffect(() => getPeople(), []);
+    const handleLogout = () => {
+        setPeople([]);
+    }
+
+    useEffect(() => {
+        const getPeople = async () => {
+            if(!props.user) return;
+            const token = await props.user.getIdToken();
+    
+            const response = await fetch(URL, {
+                method: 'GET',
+                headers: {
+                    'Authorization': 'Bearer ' + token
+                }
+            });
+    
+            const data = await response.json();
+    
+            setPeople(data);
+        }
+        if(props.user) {
+            getPeople()
+        } else {
+            handleLogout()
+        }
+    }, [props.user]);
 
     return (
         <main>
